requirements.txt
anthropic>=0.40.0
openai>=1.50.0
google-genai>=0.8.0
pydantic-settings>=2.0.0
rich>=13.0.0
